refactor: refactor and add packages to use more oop
* move commands and parsing logic to 'commands' package
* move exceptions to 'exceptions' package
* move tasks and task-related logic to 'tasks' package
* move storage and UI-related classes to 'storage' and 'ui' packages
* update file paths in tests and UI resources to reflect the new structure
* improve overall modularity and adherence to OOP principles
