+++ docs/README.md
@@ -1,4 +1,4 @@
-# EchoBot User Guide
+# echo.EchoBot User Guide
 
 // Update the title above to match the actual product name
