+++ src/main/java/Main.java
@@ -1,10 +1,12 @@
 import java.io.IOException;
 
+import echo.EchoBot;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
 import javafx.scene.layout.AnchorPane;
 import javafx.stage.Stage;
+import ui.MainWindow;
 
 /**
  * A GUI for Duke using FXML.
