+++ src/main/java/commands/Parser.java
@@ -1,5 +1,9 @@
-package echo;
+package commands;
 
+
+import exceptions.DukeException;
+import tasks.Task;
+import tasks.TaskList;
 
 /**
  * The Parser class is responsible for interpreting user commands
