+++ src/main/java/tasks/Deadline.java
@@ -1,4 +1,4 @@
-package echo;
+package tasks;
 
 import java.io.Serializable;
 import java.time.LocalDateTime;
