+++ src/main/java/tasks/Event.java
@@ -1,4 +1,4 @@
-package echo;
+package tasks;
 
 import java.io.Serializable;
