+++ src/main/java/tasks/TaskType.java
@@ -1,4 +1,4 @@
-package echo;
+package tasks;
 
 /**
  * Represents the different types of tasks that can be
