+++ src/main/java/ui/DialogBox.java
@@ -1,3 +1,5 @@
+package ui;
+
 import java.io.IOException;
 import java.util.Collections;
