+++ src/main/java/ui/MainWindow.java
@@ -1,3 +1,6 @@
+package ui;
+
+import echo.EchoBot;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
 import javafx.scene.control.ScrollPane;
