+++ text-ui-test/EXPECTED.TXT
@@ -6,7 +6,7 @@ Hello from
 |____/ \__,_|_|\_\___|
 
 ____________________________________________________________
-Hello! I'm EchoBot
+Hello! I'm echo.EchoBot
 What can I do for you?
 ____________________________________________________________
 ____________________________________________________________
